Change the type of `visual_line_count`'s `width` argument to `usize`

+++ src/multi.rs
@@ -271,7 +271,7 @@ impl MultiState {
         }
 
         let width = match self.width() {
-            Some(width) => width as f64,
+            Some(width) => width as usize,
             None => return Ok(()),
         };
